refactor(tests): introduce utils.py
